backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic[email]
passlib
python-jose
alembic
openai
openai-guardrails
openai-agents